Reduce false positives/negatives
## An attacker can bypass this rule by:
* not using carriage returns (ending commands with ```|0A|```, instead of ```|0D 0A|```)
* using a lower case "auth"
I have a POC that uses the above to crash a fork for EXIM without alerting.

## False positives:
* there should only be 3 white space separated words on the line (```AUTH PLAIN BBBBCCC D``` will not cause an overflow)
* Including equals signs in the base64 will not cause an overflow

An off-by-one overflow has no room for error. Spaces and equals signs are allowed in the base64 algorithm. They will be included in the allocation math, but won't result in any writes. So an overflow will not occur.